+++ contrib/dopus/Program/main10.c
@@ -427,10 +427,10 @@ int *count;
 
     Forbid();
     for (a=0;;a++) {
-        lsprintf(buffer,"%s%ld",config_replyport_basename,a);
+        lsprintf(buffer,"%s%ld",config_replyport_basename,(long)a);
         if (!(FindPort(buffer)))
          {
-          lsprintf(buffer,"%s.%ld",base,a+1);
+          lsprintf(buffer,"%s.%ld",base,(long)(a+1));
           if (!(FindPort(buffer)))
            {
             port=LCreatePort(buffer,0);
@@ -453,9 +453,9 @@ int win,unit;
     wsave=main_proc->pr_WindowPtr;
     main_proc->pr_WindowPtr=(APTR)-1;
 
-    lsprintf(buf,"DF%ld:",unit);
+    lsprintf(buf,"DF%ld:",(long)unit);
     if (!(lock=Lock(buf,ACCESS_READ))) {
-        lsprintf(buf,"PC%ld:",unit);
+        lsprintf(buf,"PC%ld:",(long)unit);
         if (!(lock=Lock(buf,ACCESS_READ))) return(0);
     }
     UnLock(lock);
