+++ contrib/dopus/Program/main3.c
@@ -304,7 +304,7 @@ UWORD ownerid,groupid;
               strcpy(newentry->dispstr,dispstr);
 //            else return(0);
         }
-        lsprintf(newentry->name,"%ld",dir->total);
+        lsprintf(newentry->name,"%ld",(long)dir->total);
     }
     else {
         LStrnCpy(newentry->name,name,FILEBUF_SIZE-2);
