+++ contrib/dopus/Program/view.c
@@ -204,7 +204,7 @@ void __saveds view_file_process()
 
   Forbid();
   for (a=0;;a++) {
-    lsprintf(portname,"DOPUS_VIEW%ld",a);
+    lsprintf(portname,"DOPUS_VIEW%ld",(long)a);
     if (!(FindPort(portname))) break;
   }
   Permit();
@@ -371,7 +371,7 @@ D(bug("linecount = %ld\n",vdata->view_line_count));
                   for (a=0;a<MAXTABS;a++) view_console_unit->cu_TabStops[a]=a*config->tabsize;
                   view_console_unit->cu_TabStops[MAXTABS-1]=0xffff;
 
-                  lsprintf(buf,"\x9b\x30\x20\x70\x9b%ld\x75\x9b%ld\x74",vdata->view_max_line_length+1,vdata->view_lines_per_screen); //turn off cursor, set line length, set page height
+                  lsprintf(buf,"\x9b\x30\x20\x70\x9b%ld\x75\x9b%ld\x74",(long)vdata->view_max_line_length+1,(long)vdata->view_lines_per_screen); //turn off cursor, set line length, set page height
                   view_print(vdata,buf,1,strlen(buf));
 
                   vdata->view_max_line_length=255;
@@ -1892,8 +1892,12 @@ int line;
     buf3[8],buf3[9],buf3[10],buf3[11],buf3[12],buf3[13],buf3[14],buf3[15],
     buf2);*/
   lsprintf((char *)textbuf,
-    "%08lx: %08lx %08lx %08lx %08lx %s\n",line*16,
-    ((long *)hex)[0],((long *)hex)[1],((long *)hex)[2],((long *)hex)[3],buf2);
+    "%08lx: %08lx %08lx %08lx %08lx %s\n",(long)(line*16),
+    (long)((ULONG *)hex)[0],
+    (long)((ULONG *)hex)[1],
+    (long)((ULONG *)hex)[2],
+    (long)((ULONG *)hex)[3],
+    buf2);
   if (c>-1) {
     for (b=c;b<46;b++)
       textbuf[b]=' ';
